requirements.txt
asgiref==3.8.1
attrs==23.2.0
cfgv==3.4.0
distlib==0.3.8
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==25.9.2
filelock==3.15.4
identify==2.5.36
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
psycopg==3.1.19
pytest==8.2.2
pytest-django==4.8.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.3
